==> 02 - Kaggle - bike share system - data preprocessing.ipynb <==
# 02 - Kaggle - bike share system - Data preprocessing

For problem formulation refer to **"01 - Kaggle - bike share system - problem formulation.ipynb"**.

In this section we extract date and time from the `datetime` column of the dataset using the time series capabilities of pandas.

Also, we make two data sets. On keeps the numerical values of the categorical data as is and the other turns the categorical data into dummy matrices. We will later explore the performance of these two type of data representation in our machine learning analyses.

## Data set

The **training set** includes:

<center> datetime | season | holiday | workingday | weather | temp | atemp | humidity | windspeed | casual | registered | count </center>


- **datetime**: `YYYY-MM-DD HH:00:00` --> hourly date + timestamp by hour  
    - `YYYY` = 2011 or 2012
    - `MM` = 1 - 12  
    - `DD` = 1 - 19
    - `HH` = 0 - 23

- **season**: Kaggle's [website](https://www.kaggle.com/c/bike-sharing-demand/data) says "`1 = spring, 2 = summer, 3 = fall, 4 = winter`", but the season indecies in the dataset correspond to 
    - 1 = Winter (January-March)
    - 2 = Spring (April-June)
    - 3 = Summer (July-September)
    - 4 = Fall (October-December)

- **holiday**: whether the day is considered a holiday 
    - 0 = non-holiday
    - 1 = holiday

- **workingday**: whether the day is neither a weekend nor holiday
     - 0 = day is weekend or holiday
     - 1 = otherwise 
- **weather**: encoded to make explicit various extreme weather events
    - 1 = Clear, Few clouds, Partly cloudy, Partly cloudy 
    - 2 = Mist + Cloudy, Mist + Broken clouds, Mist + Few clouds, Mist
    - 3 = Light Snow, Light Rain + Thunderstorm + Scattered clouds, Light Rain + Scattered clouds)
    - 4 = Heavy Rain + Ice Pallets + Thunderstorm + Mist, Snow + Fog

- **temp**: temperature in Celsius. 

- **atemp**: "feels like" temperature in Celsius.

- **humidity**: relative humidity

- **windspeed**: wind speed

- **casual**: number of non-registered user rentals initiated

- **registered**: number of registered user rentals initiated

- **count**: number of total rentals (casual + registered)


The ** test set ** includes the same features except `casual`, `registered` and `count`. Also, `DD` in **`datetime`** is from 20 to the end of the month.

## Numerical and categorical variables:

The data set has two types of features:

  * **Categorical** variables (integer): 'datetime' (that we will split into 'hour', 'weekday', 'month', and 'year'), 'season, 'holiday', 'workingday', 'weather'.

  * **numerical** variables (float): 'temp', 'atemp', 'humidity', and 'windspeed'.

The targets, that is, `casual`, `registered`, and `count`, are integer.

Here are the operations to follow:
  * Extract the values of 'hour', 'weekday', 'month', and 'year' from the 'datetime' column.
  * Turn all the continous variables to float.
  * For categorical variables we convert the valuyes of each categorical feature into a “dummy” or “indicator” matrix. If a feature column in a DataFrame has k distinct values, we trun it into a matrix containing k columns of 1’s and 0’s. pandas has a `get_dummies` function to facilitate this process.

Then we save the dataframes in two structures: 
 * In `train_prep_orig.csv` and `test_prep_orig.csv` the categorical data are in the original form.
 * In `train_prep_dum.csv` and `test_prep_dum.csv` the categorical data are converted to dummy matrices. 


### Basic settings and importing the libraries

# Resets the namespace by removing all names defined by the user without asking for confirmation
%reset -f

# Panas is used as a DataFrame
import pandas as pd

# Used for display dataframes as html tables
from IPython.display import display

### Importing the train data and the test data

#### train set

#Load train data and setting the datetime as index
data_train = pd.read_csv('data/train.csv', index_col=0, parse_dates=True)
print "The shape of the train dataset:", data_train.shape
display(data_train.head())

# convert humidity data type to flaot
data_train.humidity = data_train.humidity.astype(float)

# Extracting hour, weekday, month and year from datetime 
# Since parse_dates=True, we can use index's attributes .hour, .weekday, .year, .month  
data_train['hour'] = data_train.index.hour
data_train['month'] = data_train.index.month
data_train['weekday'] = data_train.index.weekday
data_train['year'] = data_train.index.year

print "extract date and time"
data_train.reset_index(level=None, drop=True, inplace=True, col_level=0, col_fill='')
display(data_train.head())

print "Statistics"
display(data_train.describe()) # There is no missing data

display(data_train.dtypes)

#### test set

#Load test data and setting the datetime as index
data_test = pd.read_csv('data/test.csv', index_col=0, parse_dates=True)
print "The shape of the train dataset:", data_test.shape
display(data_test.head())

# convert humidity data type to flaot
data_test.humidity = data_test.humidity.astype(float)

# Extracting hour, weekday, month and year from datetime 
# Since parse_dates=True, we can use index's attributes .hour, .weekday, .year, .month  
data_test['hour'] = data_test.index.hour
data_test['month'] = data_test.index.month
data_test['weekday'] = data_test.index.weekday
data_test['year'] = data_test.index.year

print "extract date and time"
data_test.reset_index(level=None, drop=True, inplace=True, col_level=0, col_fill='')
display(data_test.head())


print "Statistics"
display(data_test.describe()) # There is no missing data

display(data_test.dtypes)

### Exporting `train_prep_orig.csv` and `test_prep_orig.csv` 


cat_var = ['year','season', 'month', 'weekday', 'hour', 'workingday', 'holiday', 'weather']
num_var = ['temp', 'atemp', 'humidity', 'windspeed']
target_var = ['casual', 'registered', 'count']

data_train = data_train[num_var + cat_var + target_var]
data_train.to_csv('data/train_prep_orig.csv', index=False)

data_test = data_test[num_var + cat_var]
data_test.to_csv('data/test_prep_orig.csv', index=False)

### Converting the categorica data to dummy matrices
### Exporting `train_prep_dum.csv` and `test_prep_dum.csv`

dummy_train = data_train[num_var].copy()

# to conver t eyar to a dummy matrix, we need to differentiate between 2011 and 2012, so no need to make two columns
dummy_train['is2011'] = (data_train.year==2011).map(int)

for attrib in cat_var[1:]:
    dummy_train = dummy_train.join(pd.get_dummies(data_train[attrib], prefix = 'd_'+ attrib))

dummy_train = dummy_train.join(data_train[target_var].astype(float))

dummy_test = data_test[num_var].copy()

dummy_test['is2011'] = (data_test.year==2011).map(int)

for attrib in cat_var[1:]:
    dummy_test = dummy_test.join(pd.get_dummies(data_test[attrib], prefix = 'd_'+ attrib))
    
dummy_train.to_csv('data/train_prep_dum.csv', index=False)
dummy_test.to_csv('data/test_prep_dum.csv', index=False)

print "train set"
display(dummy_train.head())
display(dummy_train.tail())

print "test set"
display(dummy_test.head())
display(dummy_test.tail())
